+++ NEWS.md
@@ -1,7 +1,8 @@
 # oysteR (development branch)
   * Feature: Extract packages from `requirements.txt` and `renv.lock` files
   * Feature: Handle more general vulnerabilities
-  * Feature: Add `audit_pkg()` function
+  * Feature: Add `audit_description()` function
+  * Feature: Add `expect_secure()` for the {testthat} package
   * Add Josiah Perry as contributor
   * Minor: Link to [https://www.jumpingrivers.com](jumpingrivers.com) blog post
   * Minor: Make `verbose = FALSE` completely silent